+++ pelican/plugins/myst_reader/myst_reader.py
@@ -1,14 +1,20 @@
 """Reader that processes MyST Markdown and returns HTML5."""
 from __future__ import annotations
 
+from copy import deepcopy
+from enum import Enum
+from io import StringIO
 import math
 import os
 from pathlib import Path
+from typing import Any, Iterable
+import warnings
 
 from bs4 import BeautifulSoup, element
 from markdown_it.renderer import RendererHTML
+from markdown_it.token import Token
 from mwc.counter import count_words_in_markdown
-from myst_parser.config import main
+from myst_parser.config.main import MdParserConfig
 from myst_parser.docutils_ import Parser as DocutilsParser
 from myst_parser.parsers.mdit import create_md_parser
 import yaml
@@ -17,8 +23,8 @@ from pelican import signals
 from pelican.readers import BaseReader
 from pelican.utils import pelican_open
 
-from ._docutils_renderer import myst2html as myst2html_with_docutils
-from ._sphinx_renderer import myst2html as myst2html_with_sphinx
+from ._docutils_renderer import docutils_renderer
+from ._sphinx_renderer import sphinx_renderer
 
 DEFAULT_READING_SPEED = 200  # Words per minute
 
@@ -42,6 +48,61 @@ FILE_EXTENSIONS = [
     "myst",
 ]
 
+# Default MyST settings common to all parsers.
+DEFAULT_MYST_SETTINGS = {
+    # Set the default list of warnings to suppress. List available at:
+    # https://myst-parser.readthedocs.io/en/latest/configuration.html#build-warnings
+    "myst_suppress_warnings": [
+        # Ignore the "Document headings start at H2, not H1 [myst.header]" warning,
+        # since Pelican will use the title field in the front-matter as the H1 header
+        # of the page.
+        "myst.header",
+    ],
+}
+
+# List of implemented renderers.
+# TODO: refactor Renderer management with classes to make code more readable and avoid
+# passing this enum as parameters everywhere. This should remove lots of duplicate code
+# and make the addition of new rendered easier.
+RENDERER = Enum("Renderer", ["DOCUTILS", "SPHINX"])
+
+# Default Docutils settings.
+# These are the same default as the one hard-coded in Pelican:
+# https://github.com/getpelican/pelican/blob/1f6b344/pelican/readers.py#L255-L262
+DEFAULT_DOCUTILS_SETTINGS = {
+    "initial_header_level": "2",
+    "syntax_highlight": "short",
+    "input_encoding": "utf-8",
+    "halt_level": 2,
+    "traceback": True,
+    "warning_stream": StringIO(),
+    "embed_stylesheet": False,
+    # Default set of MyST extensions.
+    "myst_enable_extensions": set(),
+} | DEFAULT_MYST_SETTINGS
+
+# Default Sphinx settings.
+# These are going to be transformed into a Sphinx conf.py file.
+DEFAULT_SPHINX_SETTINGS = {
+    # Dummy settings required by Sphinx. They have no influence on the produced output.
+    "project": "myst2html",
+    "author": "pelican-myst-reader",
+    # Default set of Sphinx extensions.
+    "extensions": {
+        "myst_parser",
+        "sphinx.ext.mathjax",
+        "sphinx.ext.autosectionlabel",
+    },
+    "bibtex_bibfiles": [],
+    # Forces all links to be external
+    "myst_all_links_external": True,
+    # Default set of MyST extensions.
+    "myst_enable_extensions": {
+        "colon_fence",
+        "deflist",
+    },
+} | DEFAULT_MYST_SETTINGS
+
 
 class MySTReader(BaseReader):
     """Convert files written in MyST Markdown to HTML 5."""
@@ -50,22 +111,77 @@ class MySTReader(BaseReader):
     file_extensions = FILE_EXTENSIONS
 
     def __init__(self, *args, **kwargs):
+        """Fetch settings from ``pelicanconf.py`` and initialize parsers."""
         super().__init__(*args, **kwargs)
 
-        # Get settings set in pelicanconf.py
-        extensions = self.settings.get("MYST_EXTENSIONS", [])
-        self.parser_config = main.MdParserConfig(
-            # renderer="docutils",
-            enable_extensions=extensions
+        # Merge user-defined settings with defaults.
+        self.docutils_settings = deepcopy(
+            DEFAULT_DOCUTILS_SETTINGS
+        ) | self.settings.get("MYST_DOCUTILS_SETTINGS", dict())
+        self.sphinx_settings = deepcopy(DEFAULT_SPHINX_SETTINGS) | self.settings.get(
+            "MYST_SPHINX_SETTINGS", dict()
         )
-        self.md_parser = create_md_parser(self.parser_config, RendererHTML)
-        self.myst_extensions = self.parser_config.enable_extensions
 
+        # Add user-activated MyST extensions to the defaults.
+        myst_extensions = self.settings.get("MYST_EXTENSIONS", set())
+        if myst_extensions:
+            warnings.warn(
+                "MYST_EXTENSIONS will soon be deprecated. Use "
+                "MYST_DOCUTILS_SETTINGS['myst_enable_extensions'] and "
+                "MYST_SPHINX_SETTINGS['myst_enable_extensions'] instead.",
+                FutureWarning,
+            )
+            self.docutils_settings["myst_enable_extensions"].update(myst_extensions)
+            self.sphinx_settings["myst_enable_extensions"].update(myst_extensions)
+
+        # Parse and validate MyST settings.
+        docutils_myst_conf, normalized_setting = self._validate_myst_settings(
+            self.docutils_settings
+        )
+        # Reintegrate normalized settings to the renderer settings.
+        self.docutils_settings |= normalized_setting
+
+        # Parse and validate MyST settings.
+        sphinx_myst_conf, normalized_setting = self._validate_myst_settings(
+            self.sphinx_settings
+        )
+        # Reintegrate normalized settings to the renderer settings.
+        self.sphinx_settings |= normalized_setting
+
+        # Create a MyST parser for each renderer with its own config.
+        self.docutils_myst_parser = create_md_parser(docutils_myst_conf, RendererHTML)
+        self.sphinx_myst_parser = create_md_parser(sphinx_myst_conf, RendererHTML)
+
+        # Create a Docutils parser once to not have to re-create it for each file.
         self.force_sphinx = self.settings.get("MYST_FORCE_SPHINX", False)
+        # Save the creation of a parser if Sphinx is forced.
         if not self.force_sphinx:
-            self.parser = DocutilsParser()
+            self.docutils_parser = DocutilsParser()
 
-    def read(self, source_path):
+    def _validate_myst_settings(
+        self, settings: dict[str, Any]
+    ) -> tuple(MdParserConfig, dict[str, Any]):
+        """Parse, validate and normalize MyST settings.
+
+        Returns a MyST parser configuration object and a dictionary of normalized MyST
+        settings to be re-integrated to renderer settings.
+        """
+        # Extract MyST settings from the settings.
+        myst_settings = {
+            param_id.split("myst_", 1)[1]: param_value
+            for param_id, param_value in settings.items()
+            if param_id.startswith("myst_")
+        }
+
+        myst_config = MdParserConfig(**myst_settings)
+
+        normalized_setting = {
+            f"myst_{p_id}": p_value for p_id, p_value in myst_config.as_dict().items()
+        }
+
+        return myst_config, normalized_setting
+
+    def read(self, source_path: str) -> tuple[str, dict[str, Any]]:
         """Parse MyST Markdown and return HTML5 markup and metadata."""
         # Get the user-defined path to the MyST executable or fall back to default
         # Open Markdown file and read content
@@ -73,13 +189,15 @@ class MySTReader(BaseReader):
         with pelican_open(source_path) as file_content:
             content = file_content
 
-        # Retrieve HTML content and metadata
-        metadata = self._extract_metadata(content)
-        output = self._create_html(source_path, content)
+        # Retrieve HTML content and the renderer used.
+        output, renderer = self._create_html(source_path, content)
+
+        # Retrieve metadata with the same configuration as the renderer.
+        metadata = self._extract_metadata(content, renderer)
 
         return output, metadata
 
-    def _create_html(self, source_path, content):
+    def _create_html(self, source_path: str, content: str) -> tuple[str, RENDERER]:
         """Create HTML5 content."""
 
         # Find and add bibliography if citations are specified
@@ -89,7 +207,7 @@ class MySTReader(BaseReader):
             bib_files = ()
 
         stem = Path(source_path).stem
-        output = self._run_myst_to_html(
+        output, renderer = self._run_myst_to_html(
             content, bib_files=bib_files, tempdir_suffix=stem
         )
 
@@ -99,9 +217,9 @@ class MySTReader(BaseReader):
         for encoded_str, raw_str in ENCODED_LINKS_TO_RAW_LINKS_MAP.items():
             output = output.replace(encoded_str, raw_str)
 
-        return output
+        return output, renderer
 
-    def _calculate_reading_time(self, content):
+    def _calculate_reading_time(self, content: str) -> str:
         """Calculate time taken to read content."""
         reading_speed = self.settings.get("READING_SPEED", DEFAULT_READING_SPEED)
         wordcount = count_words_in_markdown(content)
@@ -119,7 +237,7 @@ class MySTReader(BaseReader):
 
         return reading_time
 
-    def _process_metadata(self, myst_metadata):
+    def _process_metadata(self, myst_metadata: dict[str, Any]) -> dict[str, Any]:
         """Process MyST metadata and add it to Pelican."""
         formatted_fields = self.settings["FORMATTED_FIELDS"]
 
@@ -136,12 +254,12 @@ class MySTReader(BaseReader):
 
             if key in formatted_fields and isinstance(p_value, str):
                 # Convert metadata values in markdown, if any: for example summary
-                metadata[key] = self._run_myst_to_html(p_value)
+                metadata[key], _ = self._run_myst_to_html(p_value)
 
         return metadata
 
     @staticmethod
-    def _extract_contents(html_output):
+    def _extract_contents(html_output: str) -> str:
         """Extracts contents inside a <main> ... </main> tag"""
         soup = BeautifulSoup(html_output, "html.parser")
         main = soup.find("main")
@@ -150,9 +268,9 @@ class MySTReader(BaseReader):
             str(tag) for tag in main.children if isinstance(tag, element.Tag)
         )
 
-    def _extract_metadata(self, content):
+    def _extract_metadata(self, content: str, renderer: RENDERER) -> dict[str, Any]:
         """Extract metadata from MyST markdown content"""
-        tokens = self._run_myst_to_tokens(content)
+        tokens = self._run_myst_to_tokens(content, renderer)
 
         if not tokens:
             raise ValueError("Could not find metadata. File is empty.")
@@ -188,32 +306,60 @@ class MySTReader(BaseReader):
             )
         return metadata
 
-    def _run_myst_to_tokens(self, content):
+    def _run_myst_to_tokens(self, content: str, renderer: RENDERER) -> list[Token]:
         """Execute the MyST parser and generate the syntax tree / tokens"""
-        return self.md_parser.parse(content)
+        myst_parser = (
+            self.sphinx_myst_parser
+            if renderer == RENDERER.SPHINX
+            else self.docutils_myst_parser
+        )
+        return myst_parser.parse(content)
 
-    def _run_myst_to_html(self, content, bib_files=None, tempdir_suffix=None) -> str:
-        """Execute the MyST parser and return output."""
+    def _run_myst_to_html(
+        self,
+        content: str,
+        bib_files: Iterable[str | Path] | None = None,
+        tempdir_suffix: str | None = None,
+    ) -> tuple(str, RENDERER):
+        """Select the right MyST renderer for each file and return output.
+
+        The Sphinx renderer is automatically used if:
+
+        - any math extension is enabled, or
+        - BibTeX files are found, or
+        - user's settings force the use of Sphinx.
+        """
         if (
             self.force_sphinx
             or bib_files
-            or any(ext in ("dollarmath", "amsmath") for ext in self.myst_extensions)
+            or self.sphinx_settings["myst_enable_extensions"].intersection(
+                ("dollarmath", "amsmath")
+            )
             or any(
                 syntax in content for syntax in ("{filename}", "{static}", "{attach}")
             )
         ):
-            return myst2html_with_sphinx(
-                content,
-                bib_files=bib_files,
-                myst_extensions=self.myst_extensions,
-                sphinx_extensions=["sphinx.ext.autosectionlabel"],
-                tempdir_suffix=tempdir_suffix,
+            return (
+                sphinx_renderer(
+                    content,
+                    conf=self.sphinx_settings,
+                    bib_files=bib_files,
+                    tempdir_suffix=tempdir_suffix,
+                ),
+                RENDERER.SPHINX,
             )
         else:
-            return myst2html_with_docutils(content, self.myst_extensions, self.parser)
+            return (
+                docutils_renderer(
+                    content,
+                    conf=self.docutils_settings,
+                    parser=self.docutils_parser,
+                ),
+                RENDERER.DOCUTILS,
+            )
 
     @staticmethod
-    def _find_bibs(source_path):
+    def _find_bibs(source_path: str) -> list[str]:
         """Find bibliographies recursively in the sourcepath given."""
         bib_files = []
         filename = os.path.splitext(os.path.basename(source_path))[0]